test: add search tests
